refactor: 영화 데이터 쿼리 키 관련, contants 폴더 트리변경
